refactor: rename filter_contests to contests
